free done ?

+++ sources/raycaster/raycast_cal.c
@@ -6,7 +6,7 @@
 /*   By: jbrousse <jbrousse@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/25 14:50:15 by monsieurc         #+#    #+#             */
-/*   Updated: 2024/05/27 11:25:48 by jbrousse         ###   ########.fr       */
+/*   Updated: 2024/05/28 15:48:35 by jbrousse         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,11 +16,11 @@ t_vector2	get_coord_texture(t_ray *ray, t_data *data)
 {
 	t_vector2	texture;
 
-	texture.x = (int)(ray->wall_x * data->texture[0]->size.x);
+	texture.x = (ray->wall_x * data->texture[0]->size.x);
 	if ((ray->hit_axis == 0 && ray->dir.x > 0) || (ray->hit_axis == 1 && ray->dir.y < 0))
-		texture.x = (int)data->texture[0]->size.x - texture.x;
+		texture.x = data->texture[0]->size.x - texture.x;
 	ray->step_texture_wall = 1.0 * data->texture[0]->size.y / ray->line_height;
-	texture.y = (int)(ray->draw_start -( WIN_HEIGHT / 2) + (ray->line_height / 2)) * ray->step_texture_wall;
+	texture.y = (ray->draw_start -( WIN_HEIGHT / 2) + (ray->line_height / 2)) * ray->step_texture_wall;
 	return (texture);
 }
